requirements.txt
django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
pyrebase4==4.7.1
rabbitmq_utils
stripe==7.9.0